New stratergy involves exporting instances of storybook for downstream use
