+++ package.json
@@ -13,17 +13,18 @@
     "packages/*"
   ],
   "dependencies": {
-    "antd": "^3.2.0",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
-    "@storybook/react": "^4",
-    "@storybook/addon-actions": "^4"
+
   },
   "devDependencies": {
     "@babel/core": "^7",
     "husky": "1.2.0",
     "lerna": "^3.4.3",
-    "lint-staged": "^8.1.0"
+    "lint-staged": "^8.1.0",
+    "antd": "^3.2.0",
+    "react": "^16.2.0",
+    "react-dom": "^16.2.0",
+    "@storybook/react": "^4",
+    "@storybook/addon-actions": "^4"
   },
   "husky": {
     "hooks": {
